LPS-67003 Use queries instead of model in MediaSource
We need to query both by model (fileVersion in this case) and by
properties.

To allow that, and also let other implementations use a custom DSL to
construct queries, we keep the concrete type of the qury builder
implementation dependant. This means consumers of the API must know the
kind of media they want by requesting a particular MediaSource
implementation, which seems reasonable.

Also, it is important to mention that the query engine implementation
for adaptive images will filter by model (fileVersion), and sort by
property value. So, the result of the media source will be a stream of
all matching results, orderer by specifity (i.e. how well they match the
input query).