LPS-69589 Try to create SocialActivity even if there are conflicts
The SocialActivity model has a unique restriction that doesn't allow a
user to create two activities in the same instant of time. Tweet dates
have only a precission of minutes, so it is very easy for it to fail.

Try to create the tweet anyway increasing the tweet time, imposing a
reasonable limit so that we don't loop for too long.