LPS-70693 Override D&M ImageProcessor
We're not implementing most of the methods, as they shouldn't be
called anymore.

To avoid copy&paste, we're reusing ImageProcessorImpl#getThumbnailType
in our implementation. Another thing we'll need to resolve is the fact
that the set of supported mime types is all over the place (search for
any of them in the Adaptive Media project).