LPS-72715 Using a cloned PortletDataContext inside the callback
The PortletDataContext object is reused among import calls on different
PortletDataHandlers, changing some of its contents along the way. When
the callback is run (after all the imports have been performed) the
original PortletDataContext no longer contains the data for the ratings
portlet and causes a silent failure. Cloning it at the time doImportData
is called fixes the issue.