LPS-73184 created two isolated forms for renaming/moving wiki pages
Now wiki pages can be renamed both using the "rename" button and
pressing enter after writing a new name.

The problem essentially was a conflict between the to tabs presented on
the form: one of them should perform the page renaming and the other one
should change the page's parent.

I have separated the behaviour in two different forms, so now there is
no conflict on form sending. This new behaviour is totally transparent
for the backend logic, as the same callbacks are being called.