LPS-73018 Move class to new package to avoid having to update build.gradles
Unfortunately, because of the current semantic versioning and subrepo rules, it is nearly impossible to add a new class to a package that contains a non-ProviderType interface. This is because any class that implements such an interface will result in that module's build.gradleneeding to be updated. In this case, adding the class inside the com.liferay.layouts.admin.kernel.util package resulted in a build.gradle in the journal subrepo needing to be updated, and there seems to be no possible procedure to follow in this case, since committing this change to master will break the journal subrepo. So I moved the class to a new com.liferay.layouts.admin.kernel.model package, and renamed the class to make it more consistent with the new package.