LPS-73681 These methods are being run in the wrong order
The upgradeCalendarResourceUserIds modifies the CalendarResource table so that the userIds of the default user get replaced with the userIds of the admin user for any group CalendarResource. The updateCalendarUserIds queries the Calendar table joined with the CalendarResource table for group CalendarResources with the userId of the default user, and then updates the Calendar entries returned by the query. If this query is run after updateCalendarResourceUserIds is run, it will return no results since the CalendarResource table would no longer have any group CalendarResources with the userId of the default user. Therefore, upgradeCalendarIds needs to happen before upgradeCalendarResourceIds.