LPS-73684 Remove files/shortcuts when removing repository
Note that we need to do this last. To guarantee that things are
deleted in the correct order we need to first delete all reachable
folders and file entries depth first (this is done a couple lines
before the things I changed). Then, we remove any remaining files and
shortcuts with these new methods (those would be the ones in the root
folder and orphaned (trashed) entries).