LPS-74730 / LPS-73731 Extract a base class from workflow dynamic includes.
Note that it is not possible to extract the setServletContext() method. This
method should be overriden by Kaleo Designer's dynamic include, and in this
case there would be two protected methods setting the same reference. This is
forbidden in OSGi, so the method should indeed be implemented in every dynamic
include that does not need to have its servlet context set overriden.