LPS-76940 Using fileEntry uuid for the path
The fileVersion was being retrived by the old id which is no longer
valid after removing the original image.

Since fileVersions do not keep their uuids when being exported, it is
not safe to use it to generate the file path inside the lar. Now the
fileEntry uuid will be used to generate said path.