LPS-76595 Reconstruct file name from parent link and title
For images, the "Path" attribute does not point to the item itself,
but to a List (webpage) that will be used to display the item. So, we
need to reconstruct the location of the original file from the
parentLink, title and item extension. Additionally, the parentLink
will not point to the actual parent when the item is in the root
folder; in that case, it will contain the URL of the main webpage.