LPS-77312 Return object for more consistent method signature
While access to the data object is not needed for the base Aggregator
test case, it is needed for the base Anonymizer test case, so the base
Aggregator test case's method signature is altered here for consistency.
Also, it is more consistent in general with addSomething() methods to
return the object that was added.