LPS-80341 Extend ConfigurationFormRenderer to provide custom reads of the params so that the inputs can have any desired name
Note 1: It was not possible to move the default behavior to a DefaultConfigurationFormRenderer because DDM requires too many parameters such as ConfigurationModel which is and should remain internal. Because of this I've had to rely on "null" as a special value to trigger using DDM to read the form values. It's not super pretty, but I couldn't think of a better option.

Note 2: I've added the conversion from Map to Dictionary to avoid propagating the usage of the legacy Dictionary class to the interface. We are internally using due to OSGi's APIs, be we should not expose this undesired aspect propagate further. The cost of the conversion is small since this is only done on configuration save.