LPS-82478 portal-search-web: Put selected range in search context attribute for aggregations
Unlike Old Search Portlet, Modified Facet Portlet does not add the range to the
URL. This is intentional: range is a very complex and ugly data, and we are
working on improving our URLs.

However, search context factory takes the range parameters directly from the
URL (cf
https://github.com/liferay/liferay-portal/blob/master/portal-kernel/src/com/liferay/portal/kernel/search/SearchContextFactory.java#L60-L72)

Since we do not have it in the URL anymore, we have to add it manually.

https://issues.liferay.com/browse/LPS-82478