COMMERCE-314 Use project dependency to ensure resolution
Runtime dependencies for a module are listed in MANIFEST.MF in the
produced jar. Apparently whatever produces that list makes use of
the compile-time dependencies listed in build.gradle, so when we
specify 1.0, even though it compiles, it means that at runtime the
module won't start until there's a 1.x version deployed.