Revert "LPS-68220 Use a number for the version. This way we can easily do range checks when we need a branch for later versions."
Revert: bad UX and unnecessary change. It makes no sense to do version range checks, OpenSSO/OpenAM dies with 13.5.2 in 2020 and next version is AM 5 (a different closed-source product), see https://backstage.forgerock.com/knowledge/kb/article/a18529200#AM. PM must decide if we want to support the ForgeRock AM. We are even not sure it would be a part of this module.

This reverts commit dd65798c9295b897d9022dad55607c5b7ccdd1d6.