LPS-86344 Need to sanitize the title as a file name
In D&M this is handled with an error message in the edit page. Here we
cannot do this, as the user can go to Google from different
places (and we cannot add the error handlers dynamically to all
views).

Make sure the title we get from Google doesn't contain invalid
characters, so that it can be used to derive the file name.