LPS-93180 Sort by modified date by default
Previously, we used sort by title as default order.

Explicitly check for modifiedDate, and keep the default case separate.  Although
one case is duplicated, it serves as a form of embedded
documentation (otherwise, it may be non obvious that we support sort by
modifiedDate only by looking at the code).