LPS-93310 Fix frontend-taglib-clay yarn build issues on Windows
See https://github.com/liferay/liferay-npm-tools/pull/63 for the full
backstory, but the TL;DR is that the wrong executables were being
invoked on Windows due to problems with setting the `PATH`. This would
lead Babel v6 to be used instead of the v7 demanded by the config, and
as soon as a v7 preset was loaded the build would die.

Upgrading to v1.4.12 of liferay-npm-scripts delivers the fix.