LPS-94222 Serve via ComboServlet only in "JsFastLoad" mode
To test this, I inspected the network traffic in production mode (ie. a
portal-ext.properties file with no dev settings in it) to see the combo
mode working (everything bundled and minified). Then I restarted with a
typical dev set-up in portal-exp.properties and saw the files coming
down separately, and not minified.