LPS-94775 Remove webpack-based "watch" functionality
For this LPS we will be standardizing how we use webpack by channeling
it all through liferay-npm-scripts. First step here is to remove the two
disparate ways we're currently implementing "watch" in order to simplify
the configuration. A later step will be to re-implement "watch" in a
standard way via liferay-npm-scripts.