LPS-94775 Use webpack.config.js as file name
We are using webpack in 3 places right now, and frontend-js-web is the
only one where we use a non-standard name for the config. By renaming it
from "webpack.prod.js" to "webpack.config.js", we make it consistent
with the others and it gets picked up automatically by wepback without
an explicit `--config` switch.