LPS-94775 Don't contaminate window object with oDataParser
Instead, configure webpack to create a CommonJS module, and then use the
exported function like any other module export.

We have to tell the Liferay bundler to ignore the webpack-built bundle,
because the whole point of using webpack at all here is to avoid a
circular dependency that our bundler is not currently designed to
handle.