LPS-91754 copy file instead move on checking
This change is needed because of the way the NTFS filesystem works.  When
multiple processes open a file, NTFS creates an implicit shared lock to prevent
any process from deleting the file until all consumers release it.

This affects D&M when doing a combined update and checkin operation.  The update
operation will spawn a bunch of background OS processes (to extract metadata,
generate previews, etc.), and immediately after that it will perform the
checkin.

Part of the checkin operation involves the deletion of the PWC (precisely the
code that we're changing).  As the background processes haven't finished with
the processing yet (it usually takes more time than the checkin) this operation
fails, as those processes are working with the PWC.

Keeping the PWC around has no user visible effects, other than consuming a bit
more of disk space.