LPS-95415 Keeping the same protocol in redirects
When using an external server that receives HTTPS but connects to
Liferay via HTTP the getCurrentCompleteURL returns an http URL regardless of the preferred
protocol configured in web.server.protocol property (see
https://community.liferay.com/en_GB/forums/-/message_boards/message/5662674).

getCurrentURL returns a relative URL which will use the right protocol.
Since these are internal redirects it is safe to use a relative URL.

This issue could be also solved externally by forcing redirection of all http URLs to HTTPS in the web server.