LPS-95657 Migrate .liferaynpmscriptsrc files to npmscripts.config.js
Seeing as this LPS requires the next version of liferay-npm-scripts, it
also brings in:

https://github.com/liferay/liferay-npm-tools/pull/76

which necessitates updating the config files.

A couple of the config files were duds (the ones in portal-search and
site-navigation), so I just removed them. The other three are all
straightforward: they are just the standard preset with "asset-taglib"
added as a build dependency.