LPS-95416 Forcing account selection when no credential is stored
This covers the case when the user is logged in with only 1 account and
wants to use a different one. In this case Google was automatically
using that account.

With these changes we explicitly tell Google to allow account selection
when the credentials don't exist or have been manually revoked.