LPS-95476 Remove screenChange triggering from portal-workflow-web
This seems a little strange and indirect as a way of forcing an update.
I see that the code was originally added in LPS-81302, but testing on
master shows that it doesn't actually address the requirement of that
ticket (of avoiding width changes during animation), so I am removing
it.