LPS-95476 Fix ordering while determining target
In our `init` function we check "data-target" before "href", so the
order we use here should be the same.

Discovered this bug while testing. With the old code, jQuery hid this
issue from use, but with the new code we throw an error if we get more
than one element; it turns out we were matching 126 elements(!) because
we were finding all elements with target "javascript:;". It was a
benign bug in that we would then perform a needless DOM operation on all
those elements, harming responsiveness but not actually throwing any
errors.

Example element that has both "data-target" and "href": the "+" button
in the top nav has a "data-target" of "#_some_very_long_id" and an
"href" of "javascript:;".