LPS-95476 Replace `hasClass()` with `classList.contains()`
IE has partial (and sufficient) support.

Note some unpleasantness in here because the
product-navigation-simulation-web module passes a string with multiple
classes in it via "data-open-class", and we have to match the jQuery
semantics of `hasClass()` so as not to break things (basically given a
class string "a b", it will report `true` if the element has both
classes, but only in that order (ie. it is probably doing a basic string
comparison); so we adopt "and" semantics here.

As a follow-up, I'll look at addressing the smell of using two classes
instead of one to indicate open-ness, but that will be a bigger change
so I prefer to keep it separate.