LPS-95476 Don't use jQuery to listen on `$(document)`
Instead, bind directly to the close and toggle elements. The complexity
required to do this via event delegation is not worth it.

Note the change to the ManagementToolbar module: previously, jQuery was
binding its event listener to the document, intercepting the event
before the management toolbar could get it, and calling
`preventDefault()` on it, which meant that the info panel would toggle
as expected.

With this change, things would happen in the opposite order: the toolbar
would get the event first, toggle the info panel, and then the side
navigation plugin would get it and toggle a second time, effectively
canceling out the first toggle. The fix is to remove the unnecessary
toggling in the management toolbar; in any case, the side navigation
plug-in is made to handle this kind of thing and it's the right place to
do this kind of management.