LPS-95476 Remove buggy resize handling
In the previous commit I fixed a bug in the resize handling. In this
commit I am removing it entirely (in case this proves controversial, we
can revert this change separately).

It seems that the complexity is not warranted given that some of the
logic in here doesn't make a lot of sense; for example: When resizing
across a breakpoint from desktop to not-desktop, we hide the side
nav. When resizing from not-desktop to desktop, we also hide it. I can
understand why we might want this in the former case but it doesn't
really make sense for the latter.

By deleting the code here, we get rid of a code smell in this file,
which was that we were relying on a global event bus to manage "local"
communication (because we only wanted a single resize listener).

Even with this functionality removed, we still have a responsive design
via the CSS which changes the format of the sidenav according to the
screen size, which I think preserves the real benefit at the same
time as we get rid of the buggy stuff. In any case, these right-hand
main-column sidenavs are relatively infrequent in liferay-portal, so the
visible impact of this change.