LPS-95476 Replace jQuery $toggler instances with DOM toggler elements
Note that it is important to use `getAttribute()` instead of
`href` here because in some places (eg. frontend-taglib's
management_bar_sidenav_toggler_button/page.jsp) we rely on an "href"
attribute of the form "#fragmentId" to establish the "target" element of
the sidenav. `href` will end up returning a fully qualified URL, which
would be invalid as a selector, but `getAttribute()` returns just the
fragment, which is a valid selector.