LPS-95476 Expose `SideNavigation.initialize()`
This is going to be the principal mechanism for instantiating a new
sidenav instance on demand, replacing the jQuery version of
`$(selector).sideNavigation()`.

Demonstrating the usage in frontend-js-aui-web's "item_viewer.js" and
breaking this into a separate commit because it is somewhat tricky to
find in the UI: go to (Admin) -> (Site) -> Content & Data -> Web Content
-> (Some basic web content) -> (Add image in Alloy Editor) -> (Some
item) -> (Info icon, bottom right).