LPS-95476 SideNavigation is no longer a jQuery plugin
Most of this commit is just stripping out the plugin plumbing and
hoisting the remaining actual logic it out of the wrapper function (`git
diff -w` or equivalent will show the actually relevant changes by
ignoring the indentation change). We don't need the wrapper function
because this is being transpiled as an ES6 module.

Additionally, we remove one stale reference to `&.ajax` which no longer
applies.

We still have a couple of references to jQuery (the `trigger()` and "DOM
ready" calls, for example), which I will remove shortly.