LPS-96415 Set up fallback Prettier config
Because we want to make sure that the config is applied as consistently
and as broadly as possible, this commit sets up a fallback that will
make our standard Prettier formatting config available even in a new
checkout (ie. one where we haven't done a build yet, or installed
"liferay-npm-scripts").