LPS-96466 Fix sidenav on old Chrome and old Firefox
This worked fine on latest Chrome, Firefox, and on IE 11, but it doesn't
work on old Firefox (eg. 60) or old Chrome (eg. 65). The latter in
particular is a critical problem because it's the one we use in CI.

Inspection revealed a discrepancy in the behavior of `fetch()` on those
older browsers (they weren't sending cookies, causing the network
requests to fail), which we can fix by switching to our standard fetch
utility (that we just added in 91d3054b), which includes cookies with
all requests by default.