LPS-96564 Don't use local service
Use the non local service to get the results.

To get accurate pagination numbers, we need to get the whole list of
elements first.  This should make no difference from a performance
perspective, as the service does exactly the same thing when asked
for paginated results (the filtering is performed in memmory).