LPS-96615 Use event delegation in side_navigation.es.js
This fixes a regression wherein the info panel button in the management
toolbar wouldn't work after selecting an item using a checkbox.
Selecting an item re-rendered the toolbar, blowing away the event
listeners responsible for toggling the info panel. The fix is to rely on
event delegation that works even as we change the toolbar.

I tried a few things to fix this before finally settling on the one in
this commit; for alternatives see:

https://github.com/wincent/liferay-portal/pull/34/commits

Namely:

- Autogenerate a unique "id" attribute if one isn't already present
  (doesn't work because Metal blows away the attribute on the next
  render).
- Autogenerate a unique selector based on other attributes of the
  element (doesn't work because there is no automatic way to determine
  which attributes are stable/static and which ones are dynamic).
- Autogenerate a unique selector base on "path" from an ancestor element
  (doesn't work because when we update the management toolbar we change
  the structure of the DOM, removing or inserting siblings, for
  example).

Finally, we settle on generating a unique selector based on whitelisted
attributes that we know to be stable.