LPS-96923 Make management toolbar info buttons work after nav
If we fail to tear down a subscription when navigating then you will
wind up with more than one event listener on a given info button. As
each listener fires it calls `.toggle()` on the side nav and
`.preventDefault()` on the event. If the number of listeners is even,
then the toggle operations cancel each other out and the side nav
doesn't remain open.

This commit makes us robust to such a scenario by making sure we don't
emit events that have already had their default prevented.

Separately, I am going to investigate the ManagementToolbar to see if I
can get it to properly tear down on navigation, avoiding the duplicate
event listeners in the first place.