LPS-97401 Update to liferay-npm-scripts v4.0.0
Major version bump because this one does linting and formatting
(previously it only did linting), so is technically a "breaking change"
in the sense that a run of `yarn checkFormat` that previously passed
could fail now. (Having said that, it doesn't actually fail because we
previously dealt with pre-existing lints LPS-97079).