LPS-97193 Prevent clicking on Version tab from closing sidebar
Prior to this commit, we were finding all close buttons in a given
container using the ".sidenav-close" class, and then calling
`getUniqueSelector()` redundantly. Because `getUniqueSelector()` is very
conservative about the selectors it generates (in order to avoid false
positives during event delegation), we wind up with an insufficiently
specific selector; ie. instead of something like:

    .image-viewer-sidenav .sidenav-close

we wind up with:

    #yui_patched_v3_18_1_1_1561468792477_1163 a

This leads to unwanted side effects, like clicking on tabs causing the
sidenav to close, because the selector will match all "a" tags in the
container.

The fix is to use an concrete selector if supplied, and only fall back
to generating one with `getUniqueSelector()` if we have to.