LPS-96486 Add first of custom lints
We wanted to provide examples of two different kinds of custom lint
rules that run on the new ESLint-powered tooling:

- "liferay/no-it-should": This is an example of a lint rule that we can
  apply to all of our frontend code, inside and outside of
  liferay-portal (eg. in the other GitHub hosted projects). As per the
  guidance here:

  https://github.com/liferay/liferay-frontend-guidelines/blob/master/guidelines/general/testing.md

  this enforces the recommendation to not start `it()` descriptions with
  the word "should", and instead use a verb.

  To keep all the lints green, I updated all the places identified by
  the lint, and updated the corresponding snapshot files, where
  applicable, to keep the tests passing.

- "liferay-portal/side-navigation": This is an example of a lint that is
  specific to liferay-portal; it guards against the use of the legacy
  jQuery-based `sideNavigation()` API and tells people to use
  `Liferay.SideNavigation()` instead.

With these two examples we now have a pattern that we can replicate for
all the other frontend things we want to lint inside and outside of this
repo.

The rules themselves live in eslint-config-liferay, which we apply via
liferay-npm-scripts. So this commit updates the dependency and changes
the configuration in one place (segments/segments-web) so that it knows
about the new liferay-portal-specific lint preset.