LPS-98111 Fix tests
Update to liferay-npm-scripts v4.8.0, which fixes the test failures
caused by the move of `debounce()` to frontend-js-web. Basically, before
this version, we weren't able to do cross-project imports in the test
suite, and now we can.

In the process of preparing the update I was able to standardize things
a little and remove some redundant config as a result.

I updated the workspace globs in the top-level "package.json" because
liferay-npm-scripts uses those globs to produce the mapping that allows
us to do cross-project imports: we don't actually need or want an
arbitrarily deep match there.

Also tested against "master-private" so will prepare a corresponding PR
for that.