LPS-97681 Restore missing JSDoc
We shouldn't remove the `@param` JSDoc here because these classes have
abstract methods that are intended to be overridden. If you look at an
overriding class like `MapGoogleMaps.es.js` you'll see that it uses
`@inheritDoc`.

So, restore the `@param`, and restore the function signatures too so
they match the docs. This requires local "disable"/"enable" annotations,
which are "ugly", but it's better to be correct than beautiful, and this
whole "abstract-method-then-override" pattern is very rare in JS anyway,
so we won't need to do this often.