LPS-97586 Update liferay-npm-scripts to v4.9.0
Two changes of note here:

1. We've bundled some common testing libraries with liferay-npm-scripts
   that means we can use them across all of our React projects. At the
   same time we updated from the versions of the libraries that we were
   already using in segments-web.

2. We turn automated source formatting back on for frontend files.

Tested like this:

On this branch, see that `ant format-source-current-branch` only checks
the changed files:

     [exec] > Task :yarnFormatModules
     [exec] yarn run v1.13.0
     [exec] $ liferay-npm-scripts fix --production false
     [exec] Prettier checked 21 files, found 0 files with problems
     [exec]
     [exec] ESLint checked 20 files, found 0 errors, found 0 warnings
     [exec] Done in 6.27s.
     [exec]
     [exec] BUILD SUCCESSFUL in 10s
     [exec] 1 actionable task: 1 executed

See that `ant format-source-all` checks everything:

     [exec] $ liferay-npm-scripts fix --production false
     [exec] Prettier checked 1450 files, found 0 files with problems
     [exec]
     [exec] /Users/greghurrell/code/portal/liferay-portal/modules/apps/hello-soy/hello-soy-navigation-web/src/main/resources/META-INF/resources/Navigation.es.js
     [exec]   18:8  warning  'Footer' is defined but never used  no-unused-vars
     [exec]   19:8  warning  'Header' is defined but never used  no-unused-vars
     [exec]
     [exec] /Users/greghurrell/code/portal/liferay-portal/modules/apps/hello-soy/hello-soy-web/src/main/resources/META-INF/resources/View.es.js
     [exec]   18:8  warning  'Footer' is defined but never used  no-unused-vars
     [exec]   19:8  warning  'Header' is defined but never used  no-unused-vars
     [exec]
     [exec] ✖ 4 problems (0 errors, 4 warnings)
     [exec]
     [exec] ESLint checked 551 files, found 0 errors, found 4 warnings
     [exec] Done in 24.10s.

Note that there is a file on "master" which currently has a formatting
error, so expect CI to complain about that. The fix is currently at:

https://github.com/brianchandotcom/liferay-portal/pull/75689

Also, the fix for those last 4 (non-fatal) warnings is at:

https://github.com/brianchandotcom/liferay-portal/pull/75696

How the partial formatting works: we pass an environment variable
containing the upstream branch name (called "git.working.branch.name" in
portal-impl/build.xml) so that `liferay-npm-scripts` can compute the
merge base and get the changed file list. I initially was going to take
the file list in via an environment variable, but I suspect that the
maximum size limits on Windows (a little over 8k characters) would be
insufficient in this repo where we have many files and the average path
length is 120 characters, which would mean we could run out of space
after only around 68 files). The alternative version is described in
detail at:

https://github.com/liferay/liferay-npm-tools/pull/165