LPS-97173 Don't inject reference
There are two implementations of the ImageProcessor, one is an OSGi
component (Adaptive Media), but the other is instantiated through
reflection (see dl.file.entry.processors property in portal.properties).

As adaptive media is not essential to this tests, using the registry will allow
it to run even if AM hasn't started yet (or it has been disabled for whatever
reason).