LPS-99399 Provide uniform way of instantiating React portlets
This addresses:

- Memory leaks due to failure to unmount React components when portlets
  are destroyed.
- Boilerplate code needed by almost every React app that uses Clay
  (propagating the spritemap via context).

Note that this introduces a circular dependency because
frontend-js-react-web now depends on frontend-taglib-clay (for the
context), and frontend-taglib-clay obviously depends on React;
fortunately, we recently added support for circular deps to the loader.

Next steps: I'll be adding a lint to ensure that people use the new API
instead of using `ReactDOM.render`. Beyond that, we can explore
providing a taglib to make instantiating React apps from JSP easier.