LPS-100825 Reset all things that depend on the ThemeDisplay/PortletRequest
The AbsoluteRedirectsFilter initializes and pushes into a thread local a new
instance of ServiceContext.  The problem is that this filter runs before
ServicePreAction is executed, so anything depending on the ThemeDisplay will
have the default values.  This was done on purpose in LPS-15293, so that the
FriendlyURLServlet can rely on the service context.

As this instance will be reused (it is cloned and partially updated in the
ServiceContextFactory class), the values depending on the ThemeDisplay (like the
plid) will never have the correct value.

Here, we reset all fields depending on the themeDisplay/portletRequet to make
sure everything is initialized properly.