LPS-100896 Add common hooks to frontend-js-react-web
There are some hooks that are so basic, common, and useful that we
want to provide them in frontend-js-react-web. Some initial
candidates I am thinking of:

- `usePrevious` (substitute for componentDidUpdate)
- `useMounted` (useful in async callbacks, so that you can check whether
  you are still mounted before firing off a set-state call)
- `useEventListener` (encapsulates common pattern of adding an event
  listener on mount and removing it on unmount)

We want to keep this list small and conservative to prevent it from
becoming an infinite grab bag of hooks, but there are definitely some
hooks (like the ones above) that are pretty uncontroversial.