LPS-102323 Save changes of JavaScript Loader makes the System Setting unavailable
This error was caused because modifying the settings of the loader
caused the NPMRegistry to reload and that caused a Loader call from the
browser to fail.

In this commit I'm avoiding the reload by providing a @Modified handler
in the component and I'm also optimizing the Tracker reopen (this is not
strictly necessary to fix the bug but should add stability and
performance to the system).