LPS-102503 Remove liferay-theme-es2015-hook
The Porygon theme is the only user of this module and we can rewrite the
JS fairly trivially to work without it, which allows us to clean up some
cruft from the top-level package.json and yarn.lock.

We move the contents of top_search.es.js into main.js with as few
changes as possible (just the necessary ones to drop the dependencies on
the metal modules).