LPS-102599 Create getEditorReactContext()
The soy version is divided in two JSP files that each need specific
data. The React version is conceptually a single app with a single set
of data requirements even though it is rendered in different places in
the UI.

So, make a new `getEditorReactContext()` which is the combination of the
two soy contexts (`getEditorSoyContext()` and
`getFragmentsEditorToolbarSoyContext()`).

In theory, the properties in these two contexts are all compatible,
except for the "portletNamespace", which is different in each. So, I
made an accompanying edit to the React toolbar template to make sure it
always uses the namespace from the context and not the one it would get
from the `<portlet:namespace/>` taglib, because we need it to be stable
and something that we can read from the React side.