LPS-102599 Add sidebar and initial CSS structure
Adds two buttons to the sidebar which currently don't do anything (I'll
wire those up to lazily load the code from the corresponding plugins
later on).

This may not be the final form that the CSS takes but it is a step
towards it: we want to colocate CSS with our components, which you can
see me doing in this commit; we need to evaluate whether we want to
actually use loading via `import` statements in conjunction with the
bundler or not.

Note also that I am not copying the CSS over from the Soy implementation
but starting from scratch because this is an opportunity for us to clean
things up (there may be dead CSS on the Soy side that we don't want to
bring over).

One of the first clean-up items is trying to make our shared
_variables.scss very clear: I've divided them into sections so that we
can see all the colors that are being used in the app (note I'm doing
this even when the color is only used in one place, so that we have
visibility into color usage and the opportunity to consolidate), and
likewise all the z-indexes go in one place with `$zIndex` as a prefix
(note I am just carrying over the old values here); finally, dimensions
which affect overall layout get their own section too.