LPS-102599 Extract some functionality into hooks
Still a work-in-progress, but this extracts `usePreload()` (for
loading modules on demand; it's possible there might be a better name
for this) and `usePlugins()` (for managing a registry of plugins), while
trying to keep the very sidebar-specific bits still in the sidebar (ie.
the management of which plugin is active, whether we should be closing
or opening etc).

Using `lazy` and `Suspense` now to show a spinnner, as well as providing
an error boundary that I'll wire up to allow retrying on failure.

Still needs some refactoring, but this just a step.