LPS-102599 Provide "Reload" button when errors occur
I suspect this might be improved by some further refactoring, but this
is what I got when I tested various failure modes, such as forcing a
module not to load, or throwing an error during render. I am using the
error boundary somewhat unconventionally here because it only catches
errors during render and it cannot catch errors that happen
asynchronously (which is what will happen when we have a module load
error); so that's why I have a separate `hasError` state in the parent
and I'm really just using the boundary to update that.

One way to do this without that would be to change the plugin interface
to just return a React component (then, via the magic of `lazy`, a
failure to load a component could be turned into a "render-time" error);
however, I suspect we don't want to do that because the component
interface is probably too limiting for the purposes of defining plugins
(where we will probably want imperative handles, and persistent instance
state etc).

The "Reload" button isn't really a reload as such: it just resets some
state so that the next time you try to open a panel it will do things
over. In the case of a module that failed to load due to a temporary
network failure, it might work on a subsequent try. In the less likely
case that an error is thrown during render, it still might work on a
retry is some state has changed (I tested this with a fake counter that
would throw only the first few times).