LPS-102599 Provide an App proxy object for use in plug-ins
The idea here is to avoid dependencies across OSGi modules
(ie. main app uses `Liferay.Loader.require` to load plug-ins from other
OSGi modules, and then those plugins end up `require`-ing JS from the
host app).

Rather, when the main app loads a plug-in, it passes down the kind of
arbitrary app data that the plug-in might otherwise have to require. In
the case of a React app, that particularly means the context objects:
with those, we have pretty much all we need for the plug-ins to
communicate with the rest of the app (because they can `dispatch`
actions and read from the store); as seen in previous commits, they also
have the ability to load and unload reducers of their own, and of course
they can define their own actions (whether or not other plug-ins do
anything with those actions is another question, but we should see how
far we can get without adding any cross-plugin dependencies).