LPS-103292 Add "useTimeout" and "useInterval" hooks
We need to be careful about running side effects on unmounted
components, so these hooks will help us to do the right
thing. Basically, it will replace usages of `setTimeout` and
`setInterval` with a hook that includes a baked-in `isMounted` check.

Subsequently, we will add a lint for `setTimeout` and `setInterval` in
React components and suggest people use these hooks instead.

Originally reviewed at:

https://github.com/jbalsas/liferay-portal/pull/1990