LPS-98032 Restructure JSP to avoid special cases
As requested here, we remove the special cases by restructuring the JSP.

https://github.com/brianchandotcom/liferay-portal/pull/79731#issuecomment-543483079

There were two files that we were excluding. I'll explain the simplest
one as an example:

```
{
	id: '<portlet:namespace />editAsset',
	portletAjaxable: true,

	<c:choose>
		<c:when test="<%= redirect != null %>">
			redirect: '<%= HtmlUtil.escapeJS(redirect) %>'
		</c:when>
		<c:otherwise>
			refresh: '<%= portletDisplay.getId() %>'
		</c:otherwise>
	</c:choose>
}
```

The problem there is that when the tags are temporarily removed to turn
it into JS for formatting, you're left with an invalid object (note the
lack of trailing commas):

```
{
	id: '<portlet:namespace />editAsset',
	portletAjaxable: true,
	redirect: '<%= HtmlUtil.escapeJS(redirect) %>'
	refresh: '<%= portletDisplay.getId() %>'
}
```

We can't just put a trailing comma after both properties because IE 11
will crash if the last key-value pair has a comma.

They can be rewritten — with a temporary variable — like this:

```
var data = {
	id: '<portlet:namespace />editAsset',
	portletAjaxable: true
};

<c:choose>
	<c:when test="<%= redirect != null %>">
		data.redirect = '<%= HtmlUtil.escapeJS(redirect) %>';
	</c:when>
	<c:otherwise>
		data.refresh = '<%= portletDisplay.getId() %>';
	</c:otherwise>
</c:choose>
```

which has the same behavior at runtime, but without the downside of
being invalid JS once the tags are stripped. The other example is much
more complex, so I'll let the diff speak for itself: it is basically
doing the same thing though (creating a config object and mutating it
rather than trying to do it all inline).