LPS-104970 Split config and store state inside DisplayContext
As a first cleanup of the data that is being sent to the React
editor, we are splitting config (constant values) from the actual
application state.

Moveover we are just sending the data that is being used by the
editor, so we can start adding new elements when more features
are migrated.