LPS-105364 Improve accessibility of Treeview component
We're not correctly setting the role attributes as specified here:

https://www.w3.org/WAI/GL/wiki/Using_ARIA_trees

We had role="treeview" at the top, but we weren't propagating that to
the DOM, and we were missing role="group" on nodes with children. We had
role="treeitem" on the TreeviewCard component, but that wasn't working
either because ClayCard was overwriting it with role="button".