LPS-102563 Make double-click to expand/collapse work
No idea why, but when we use `ClayCard` we never get the double-click
event; ie. instead of the expected "single", "single", "double", we just
get "single", "single". This only happens if we change the state in
response to the single clicks; ie. the sequence is:

- "single" -> toggle selected.
- "single" -> toggle selected.
- [nothing]

Instead of the desired:

- "single" -> toggle selected.
- "single" -> toggle selected.
- "double" -> toggle expanded.

Note that we aren't doing any tricks like deferring the "toggle
selected" action with timers because that would make the UI feel laggy,
it updates fast enough anyway when doing all three actions, and because
the delay needed to robustly defer is not reliably determinable across
browsers.

When I swap out `ClayCard` for identical markup, I do get the event.