LPS-105947 Remove unnecessary .babelrc
After LPS-105824, this is the only remaining use of an "ignore" rule in
a babelrc file in the repo. "ignore" rules are dangerous because they
can lead to untranspiled code making it into IE and breaking things. In
this case, we don't even need the configuration because there aren't any
"__tests__" directories anywhere in that project.