LPS-104826 Rename getContent() to saveTitle() for symmetry
We had two inverse operations here (`getContent()` and
`restoreTitle()`), but the symmetry wasn't obvious. With the new name,
the relationship between `saveTitle()` and `restoreTitle()` is clearer.

The other thing about the `getContent()` name was that it didn't sound
at all like a mutating operation; at least with `saveTitle()` it becomes
clear that we are writing data somewhere.