LPS-98114 Remove unused config from dynamic-data-mapping-form-field-type
The "main" field in this module is wrong/misleading (see LPS-98114),
but nothing is importing from it anyway (as per `git grep
'from.+dynamic-data-mapping-form-field-type'`), so we can just remove
it.

In this commit I remove the other unnecessary Jest config fields at the
same time. Tests pass before and after this change.