LPS-101720 Add LoginMVCActionCommand
We are registering this action command on the login portlet, but the
Login portlet keys are internal, so we can't reuse them.

We have also placed it in the same package as the rest of the actions,
even though it registers to a different portlet.

We checked the flow in all the methods and it seems not intertwined.
Side effects like storing attributes on the session have been pushed to
the end of the method when possible. We also avoided creating private
methods when they were not reused nor saving space.

The ordering of the liferayPortletURL parameters seems the most common
across the portal.