LPS-109467 Fix errors on reloading in journal-web
Prior to this commit, we were seeing the page blow up with errors when
you reloaded after going to: "Web Content" → "Structures" and creating a
new structure.

Specifically, incremental-dom was blowing up around here:

https://github.com/google/incremental-dom/blob/bd1415e75bdabfc452836da9ab72d9fdd9b29b48/src/assertions.ts#L214-L219

Code inspection and manually testing in the debugger revealed that this
was only happening when using the `<ClayActionsDropdown />` component,
and that the DOM structure looked like this:

- `document.body`
  - `#clay_dropdown_portal`
  - `.ddm-field-actions-container`

At the time the assertion blows up, the `currentNode` was always the
`.ddm-field-actions-container`, and the `startNode` was an object with a
`nextSibling` field pointing at the same node.

I found that by changing the DOM structure to ensure that the above
elements were not siblings under `document.body`, the error goes away;
ie. the DOM structure is now:

- `document.body`
  - `#clay_dropdown_portal`
  - `.ddm-actionable-fields-container`
    - `.ddm-field-actions-container`

I am not sure whether we're going to run into this problem again with
other instances of `ClayActionsDropdown` in this project (for example,
in `withMultiplePages.es.js`), but because this fix is somewhat
mysterious, I want to keep it as minimal as possible while still fixing
LPS-109467.