LPS-111232 Replace <liferay-ui:input-editor> with <liferay-editor:editor> in blogs-web
The goal of this change is to replace `<liferay-ui:input-editor>`
(which was deprecated in 7.1.0) with `<liferay-editor:editor>`.

A previous Jira ticket (https://issues.liferay.com/browse/LPS-110008)
and pull request (#320) are currently in review and change the same files,
so the commits on the previous PR (#320) have been cherry picked on this
branch to avoid confusion.