LPS-111374 Fix failing ddm-form js-unit tests
These tests were doubly disabled in:

- https://github.com/brianchandotcom/liferay-portal/pull/87152
- https://github.com/brianchandotcom/liferay-portal/pull/87153

Root cause:

These two modules, dynamic-data-mapping-form-builder and
dynamic-data-mapping-form-web, are the only two places left in
liferay-portal that we still use Metal's incremental-dom, which is
incompatible with React.

In f7704c187c, we switched everything to assume React by default, but
these two modules needed an explicit opt-out so that they could continue
to use incremental-dom.

All of this was fine until f18e7fe22a, which added a React-powered modal
to frontend-js-web. This caused these DDM tests to fail because they
depend on frontend-js-web, and as soon as they hit React JSX everything
blows up because without the React preset, it is considered to be a
syntax error.

The fix is to selectively mock the `openModal` export from
frontend-js-web to prevent the codepath that involves JSX from being
exercised. Consider it a nasty hack, but a temporary one, because we
will eventually get rid of all incremental-dom usage.