LPS-111979 Center CKEditor dialogs
The goal of this fix is to "center" CKEditor's dialogs.

Since our recent change to use CKEditor in more places, CKEditor dialogs
are now aligned in the top left corner instead of being centered.

The only previous commit I could find concerning this file is in
[https://github.com/liferay/liferay-portal/commit/a4df2c65a8fb39c139e0e9781078b2092c868f68](https://github.com/liferay/liferay-portal/commit/a4df2c65a8fb39c139e0e9781078b2092c868f68),
when the file was initially added.