LPS-112410 headless-admin-user-test: re-adds partial setup logic
It's important to make sure that the test user is not a member of the
test group, otherwise it will be returned in the search results for a
number of tests that do not expect it. This is outside the domain of
what DataGuard is trying to accomplish, therefore it has to stay in the
setup method.