LPS-113053 Add test
This test sets up the following scenario:
1. A calendar event is created on today's date and repeats daily until X
date
2. The fourth event in the series is deleted
3. The third event in the series is edited so that it repeats for four
days

The entire series should now start from today's date and continue for a
total of six days (inclusive). Because of the modification in Step 3,
this scenario now consists of two CalendarBookings:
1. A CalendarBooking for today's date
2. A CalendarBooking for the third event in the series

The new CalendarBooking should not share the same exceptionJCalendar
with the original CalendarBooking, or else the new CalendarBooking will
errorneously skip over the date from Step 2.