LPS-113231 Audit lints and suppressions related to arrow functions for IE
Follow-up to LPS-113193
(https://github.com/brianchandotcom/liferay-portal/pull/88445).

We added a lint rule to stop arrow functions into sneaking into
untranspiled JS that gets served to IE. This commit is result of
auditing codebase to see if there are any places where we should be
applying the rule but aren't doing so yet (found 2), and to remove any
related suppressions that are no longer required since adding the rule
(found 1; don't need it inline in frontend-theme-classic because it is
being applied to the entire module already via an .eslintrc.js file).