LPS-113021 Simplify "aui:script" tags to just "script"
We no longer have any AUI dependencies used in these scripts, so they
don't need to be "aui:script" any more.

In terms of generated output, the only difference is that the
"aui:script" tag makes an IIFE that sets up some local variables that
the scripts don't use:

```
var $ = AUI.$;var _ = AUI._;
```

`Liferay.DynamicSelect` is defined in our global bundle relatively early
on in page load, so it is already available before the code that depends
on it runs.