LPS-113569 Deprecate `liferay-language` AUI module
`Liferay.Language.get` is replaced by the language filter[1]
and we also want to remove the AUI dependency, so in this change, we
migrate the `Liferay.Language.get` function to `frontend-js-web`.

Note that the migrated version is dumbed down to just return the
key: it does not preserve the dubious fallback behavior of the old
implementation, that used a deprecated *synchronous* request to the
server to fetch the value, which would produce a console warning, and
very likely end up returning the key anyway (unless the corresponding
value happened to be in the kernel); in short, the old implementation
only wallpapered over a real problem (failure of the filter to do its
actual job).

Additional notes:

The `portal-available-languages`[2] AUI module defines 2 more functions
on the global `Liferay.Language` object.

[1]: https://github.com/liferay/liferay-portal/blob/9839677effcea2087d194c0ca29d591d5f1d8533/portal-impl/src/portal.properties#L9796-L9807)
[2]: https://github.com/liferay/liferay-portal/blob/a0bfd3552193bb33687ef3677fe4879447e04d6a/modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/available_languages.jsp