LPS-114633 Ensure CollectionDisplay is always injected
In the old sidebar panel, we where faking a "Collection Display"
component with a React component. In the new panel, this element
is being injected from withing fragmentsReducer, so no extra logic
is needed inside the Sidebar.

We want to inject this element even if the "Content Display" collection
is not present on the initial list provided by backend.