LPS-109624 Check prefers reduced motion
On certain environments like browserstack the product menu items
collapse was not working.

This was because `prefers-reduced-motion` was active and, when
that happens, CSS transitions related with collapse are set to
`transition: none`. In the CollapseProvider component we were trusting
on `transitionend` event to make everything work but, if there's no
transition, there's no way all of this could work.

With this commit we're checking if `prefers-reduced-motion` is set to
`reduce` on each `show` and `hide` request and, in that case, we avoid
all the transition stuff and we just set some css classes.

To see this in action:

- Activate reduced motion on your machine (in MacOs System Preferences >
  Accessibility > Display > Reduce motion)
- Go anywhere in portal
- Play with product menu submenus to see if they're collapsing properly.