LPS-109624 Reduce dispersion of class management code
This is a small refactor to the parent commit, which tries to keep the
class management closer together in order to make it less fragile.

Instead of an `if` conditional with an early `return`, with some classes
being set in the positive branch and different classes being set in the
`else`...

...we always set/remove the same classes and styles in the same place,
and the `if` now just decides whether to do it immediately (in the
"prefers reduced motion" case) or via a transition.

This means that the browser does some unnecessary work when "prefers
reduced motion is on" (it makes some changes and immediately undoes
them), but that is not a hot code path, users rarely use that setting,
and I think the wins in robustness make it worth it.