LPS-113561 Remove `Liferay.provide` in `account-admin-web`
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561),
we're replacing `Liferay.provide` which we eventually want to deprecate,
so here we're updating a simple usage: in this case the `addRow`
function didn't depend on any AUI modules, so we can just declare it on
the `window` object without changing the behavior.

To test this change

- From the global menu select the "Control Panel" tab, and navigate to
"Accounts".
- Click on the "+" button to create a new account
- Optionally add a name for this account
- Scroll down and find the "Valid Domains" section
- Click on the "Add" button to add a domain
- Enter a valid domain name, e.g. "liferay.com"
- Click on the "Save" button
- The "liferay.com" domain should be added to the list of "Valid
  Domains"