LPS-113561 Avoid unnecessary call to `Liferay.provide` in asset-list-web
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561),
we eventually want to deprecated the `Liferay.provide` function.

The code here was using it twice to expose functions on the `window`
object: one of those was internal only, so didn't need to be exposed,
and the other doesn't have any dependencies (which is the main reason
you would use `Liferay.provide` in the first place), so can just be
defined manually.

Additionally, we avoid the same group being added twice in the "scope"
section.

Tested as follows:

- Navigate to "Content and Data" > "Collections"
- From the "New" drop down menu, select "Dynamic Selection"
- Give it the name you want and click on the "Save" button
- Scroll down to the "Scope" section
- Click on the "Select" button
- Choose "Global"
- Click on the "Select" button again and choose "Global" again

As a result the "Global" item should be present only once in the "scope"
section.

With the devtools open, locate the input element who's `id`
attribute is `_com_liferay_asset_list_web_portlet_AssetListPortlet_groupIds` and
confirm that it's `value` attribute contains several numbers (ids), none
of which are duplicated.