LPS-113561 Avoid declaring `<portlet:namespace />selectRole` function as global
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561) we
eventually want to deprecate the `Liferay.provide` function.

In this change, we avoid declaring the `<portlet:namespace />selectRole`
function on the global `window` object and remove the unnecessary
`liferay-search-container` AUI dependency.

To test this change:

- From the "Global Menu" select the "Control Panel" tab
- Click on "Users and Organizations"
- Click on the "+" button to create a new user
- Enter a screen name (for example User01)
- Enter an email address (for example user01@liferay.com)
- Enter a value for First Name (for example User)
- Enter a value for Last Name (for example 01)
- Click on the "Save" button
- Select the "Roles" tab
- Click on the "Select" button under the "Regular Roles" section
- Choose a role in the "Select Regular Role" modal window

As a result the role you selected should be visible in the "Regular
Roles" section and no JS errors should appear in the console.