LPS-113561 Avoid declaring functions on `window` in `iframe-web`
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561) we
eventually want to deprecate the `Liferay.provide` function.

In this change, we avoid creating three functions on the global `window`
object and remove the unnecessary AUI dependencies.

To test this change:

- Navigate to "Page Builder > Pages"
- Click on the "+" button to create a new page
- Choose the "Blank" basic page template
- Enter a name for your page and click on the "Add" button
- Drag and drop the IFrame widget to your page
- Click the "kebab" menu on the right side of the Iframe widget
- Click on "Configuration"
- A modal window should open to let you configure the Iframe widget
- In the "Source URL" field enter a valid URL (for example
  https://en.wikipedia.org/wiki/Liferay)
- Click on the "Save" button
- Close the modal window
- Click on the "Publish" button
- In the left sidebar click on "Go to Site"
- Click on the page your created earlier

As a result, the Iframe widget should display the URL that you specified
in it's configuration.