LPS-115139 Make sure polyfills are loaded before used
As described in [LPS-115139](https://issues.liferay.com/browse/LPS-115139),
the call to `Object.entries` [here](https://git.io/Jfy4v) is throwing an error on IE11.

We think it shouldn't because we have a polyfill for this function in
the [frontend-compatibility-ie](https://git.io/Jfy4U) module,
but it seems that the polyfill is loaded after it is used.

So in this change, we change the location of where those scripts will be
loaded