LPS-112456 Removes unnecessary leading slashes
AbsolutePortalURLBuilderImpl.forModule already checks if the context
path ends in slash and adds one if it doesn't. It also technically is
expecting a `String relativeURL` for the module path and checks if it
starts with a slash to add it if it doesn't.

All in all, it doesn't really matter if we have a leading slash or not.

Thus, for consistency and simplicity, I've decided to get rid of them.