LPS-113561 Avoid Liferay.provide in portal-reports-engine-console-web
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561) we
eventually want to deprecate the `Liferay.provide` function, the
`<portlet:namespace />testDatabaseConnection` function is now directly declared,
on the `window` object and it's AUI dependencies have been removed.

An interesting detail about this change is that in my case the
`<portlet:namespace />testDababaseConnection` function was failing
because of a missing AUI dependency as can be seen
[here](https://git.io/Jf5Cf).

To test this change:

- Make sure you're running DXP:
  run `ant setup-profile-dxp` before running `ant all`
- Navigate to "Configuration > Reports Admin"
- Click on the "Sources" tab
- Click on the "+" button to add a new source
- Enter a name
- Enter a value for "JDBC Driver Class Name"
  (com.mysql.cj.jdbc.Driver)
- Enter a value for "JDBC URL"
  (jdbc:mysql://localhost/lportal_master?useUnicode=true&characterEncoding=UTF-8&serverTimezone=UTC)
- Enter a value for "JDBC User Name"
- Enter a value for "JDBC Password"
- Click on the "Test Database Connection" button

As a result, and depending on the values you entered in the form, you
should see a modal window appear notifying you that the connection to
the database was either successful or that it failed, and no JS errors
should appear in your browser's console.