LPS-113552 Avoid multiple DOM elements with same ID
Prior to the recent refactoring, the code in this module used to have
this behavior:

- If no container ID is provided, we create and use a default container,
  injecting it into the body.
- Subsequent toasts without container IDs will reuse the same default
  container.
- If we then provide a container (or container ID), the default
  container is ignored.

Due to refactoring, the behavior changed to:

- A div with the default container ID is always unconditionally created
  (bad).
- If no container ID is provided, we use the default container,
  injecting it into the body.
- Subsequent toasts without a container ID will inject additional copies
  if the container (bad), with identical IDs (bad).
- If we then provide a container (or container ID), we append yet
  another copy of the default container as a child, again with an
  identical ID (bad).

So, this commit fixes that:

- Default container is only created as a fallback (ie. when no valid
container or container ID is provided).
- Default container is re-used if fallback is needed multiple times (ie.
  no duplicate ID in DOM).

It preserves the change which originally motivated the refactoring:

- If a container is provided, a div is inserted into it, into which to
  mount the toasts (this is so that React won't blow away the entire
  contents of the container).

A suboptimal, but not the end of the world, consequence of this is that,
because these inserted divs don't have IDs, we don't have any way of
reusing them if multiple toasts are shown, so they will accumulate in
the DOM until the next navigation or other update removes their parents.

I renamed the `getContainerElement` function to `getRootElement` to
reduce the overloading of the word "container" just a little bit: React
uses the word "container" to describe the node into which
ReactDOM.render mounts a compenent, and Clay uses the term
`ToastContainer` for the component that contains the toast. The naming
still sucks, but at least now we can write:

    const rootElement = getRootElement(...);