LPS-115798 Allows passing an onClose handler to Liferay.Util.OpenToast
The callback will be invoked with an `{event, unmount}` argument.

The `event` property will be `undefined` if the toast is autoclosing and
of type `Event` with a reference to the close button when the toast is
manually closed.

The `unmount` property is a function that unmounts the toast and can be
invoked manually to discard it.