LPS-113561 Avoids adding event handlers when the modal is not visible
Currently, the Modal won't unmount when closing, it will simply stay
invisible. This means that the useEffect cleanup doesn't run during the
last render and we have a hanging global Liferay event waiting to run
the next time it is fired.

With this, we simlpy avoid adding the listeners (and needing to detach
them) when the modal is "closed", but the problem of the orphan components
still remains and should be addressed in a followup PR.