LPS-116752 Tweaks monospaced condition
Adjusts monospaced condition so the class is only added when the button
prints only an icon. I'm honestly not sure why we added this in the
first place, since monospaced should be manually specified rather than
implicit to prevent undesired side effects and misalignments.

Keeping it for now for backwards compatibility