LPS-117055 Setup all DDM structure forms
The new implementation of the tag assumes there is only one tag
per page, so it dumps the field values as request parameters with
no information about the structure.

As we send data about multiple structures, with the structureId as
namespace, we need to construct a request that isolates fields from
different structures, so that DDM can build the form.

Note that we cannot use NamepaceServletRequest, as DDM uses
getParameterMap internally, and that does not filter the parameters
correctly.