LPS-117924 Simplify yarn workspace globs
We occasionally get questions in Slack like on this thread...

https://liferay.slack.com/archives/CNBG06JS3/p1594970715441700

...asking why the "liferay-npm-scripts" executable can't be found at
build time. This has come up twice in the last couple of weeks alone.

The problem is usually a missing glob in the Yarn workspaces list. The
purpose of this ticket is to trial simplifying the globs so as to reduce
this type of support request, and confirm that it doesn't cause problems
in the build. Manual testing shows the build `yarn install` times after
blowing away `node_modules` are the same (41 seconds on my machine), and
additionally I made this script which compares the globs before and
after this commit:

https://gist.github.com/wincent/fea9f37abe2261526cf5b9a8191296ee

And prints the following (ie. the globs identify exactly the same set of
packages):

    Common: 127, Missing: 0, Extra: 0