LPS-117760 Visual adjustments to liferay-trash: tags
This commit is a bit of a cleanup of the liferay-trash tags that
were showing some legacy components that would not play well with
newer content and styling.

It basically removes unnecessary links that were meant to be hidden
when js was enabled to produce dual behaviour. This behaviour can
be achieved in the same way with a single submit button instead

It also makes content inline so it plays well with our alert
guidelines and removes no longer necessary css