LPS-117760 Generates title by default
Most call sites of openToast use the default type-title equivalences
so this adds a quick way to avoid that repetition.

On top of that, it forces the addition of a : character between
the title and the content as per the guidelines. This, among other
things makes the alert content more understandable when read
through an automatic screenreader