LPS-118726 Passes string to openToast message instead of a React element
Being able to directly pass a React element as a message to openToast
was never an intended API. Every other usage we have is constructing
html as a string and passing it to the openToast utility.

The current use case is simple enough; render a string or a string and
a link with an icon. In order to keep the API surface contained until a
clear pattern emerges we generate the HTML manually rather than through
a React element.