LPS-118579 message-board-web: Changed message thread declaration
The issue was that the the message thread was used before the thread was initialzed.
This caused request methods to call on a thread that was null which would throw an NPE.
Moving the point of initialization of the message thread to where the threadID is checked fixes the issue,
without having to modify other files.