LPS-118940 Trim trailing colon from toast title
This is a "defense in depth" layer on top of the parent commit. We don't
think there are any call sites which pass in a redundant colon at the
end of their `title`, but in case they do, we can gracefully suppress
it at pretty low cost.

In doing this, I also rejigged the code here to try to make it easier to
read ie.

- One ternary and one `if`, written as statements; instead of:
- Two `||`, one `&&` and a ternary, interpolated inside JSX.