LPS-118398 Expose the whole state of the translation
The importer/exporter interfaces only take/provide a view of the new changes,
not of the original ones the came from. Additionaly, there's no information
about source or target language.

For the translation preview, we need all that information, so I've created the
"TranslationSnapshot", which is (conceptually) a "picture" of the translation in
a particular state.