LPS-119479 Adds Video and Audio Selector Buttons when Xuggler is enabled
Adding this by default in the simple setup so that it matches the expectations
of admins enabling Xuggler in their system.

Product will revisit the whole audio/video integration since Xuggler is not meant
to generate production quality media and is being misused for that matter. In the
meantime, we add the buttons unconditionally for a more consistent support