LPS-118805 Styles dialog content based on .portal-popup class
We were previously using dialog-iframe-popup in an attempt to differentiate
some action-triggered dialogs from others.

Based on usage this seems rather unnecessary and causes flash of unstyled
content because the class is only added to the iframe later in time through
JS or via url params which made it complicated and hard to maintain.

By styling `portal-popup` instead, we are sure modal content is visually
consistent which improves perceived performance making content and controls
within dialogs appear ready faster.