LPS-120494 Returning true by default as below
This is needed to fix the case when there is a portlet with its own
navigation inside a display page template.

The URLs generated by the portlet (DM in our case) no longer contain any
information to identify the original info item therefore they can't be
seen.

With this change at least the page can be displayed although any
fragment mapped to the original info item will be broken.