LPS-120418 Replace removeChild implementation to avoid throwing NotFoundError in ie11
Unlike modern browser IE11 throws an error when trying to remove a child
that doesn't belong to the parent anymore. This is happening when we are
mounting dropzones. Dropzones are implemented using react portal mounted
in an html that react doesn't control and that may change without react
noticing it