LPS-121699 Back out change to tsconfig.json to satisfy SF
We have an exception to the "no trailing newline" convention only for
"package.json", and only because that's a file that gets update by other
tools (like `yarn`) which insist on putting the newline there.

https://github.com/liferay/liferay-portal/blob/8ba97db5d9cac893f46425134b84295b1dbb1ad4/modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JSONStylingCheck.java#L45-L47

The only reason this change got applied in the first place was because I
was running `yarn format` from the remote-app-client-js directory (where
the ".prettierignore" has no effect). When run from the top level,
Prettier doesn't try to change this file.