LPS-121699 Add suppression for react-hooks/exhaustive-deps rule
I am not sure whether these were pre-existing violations that snuck in,
or are new because the update version of the rule is more sensitive, but
in any case the tactic must be the same: given that the interactions
between dependencies and code can be extremely subtle, it's best if the
teams that own the code here do their own inspection and determine
whether is is safe to restructure the code. I'll ping the respective
teams.