LPS-121824 In memory of Nate Cavanaugh... please post memories of Nate to http://tinyurl.com/nate-cavanaugh as Git comments.
Dear @natecavanaugh,

I miss you bro. I recall our last conversation, the day before you died, and I
wish I had said more.

I'm so thankful to have had you as a friend and coworker the past 14 years.

Every time I look at CSS and JavaScript, or see the word Alloy/AUI, I think of
you.

Every time I format a computer, I think of you. You gave me your
[dotfiles](https://github.com/natecavanaugh/dotfiles).

Every time I merge a pull request, I think of you. I still use your
[Git tools](https://github.com/natecavanaugh/git-pull-request) and probably will
so long as GitHub and Git are a thing.

We're naming 7.4 after you and breaking our pattern of naming releases. Yeah,
you're one of the few who knows about our secret naming scheme.

I know where your body lies
[here](https://sgo.forestlawn.com/?search=1&sradio=grave&fname=Nathan&slast=Cavanaugh#sgo_results).
Your body is buried in lot 1956, space 3, in Covina Hills. But you are not your
body.

And you are not just our memories of you. Even when we all die, and everyone who
has ever met you here on Earth breathes their last, you will still be you. And
you will still be more than your body.

So Nate, where are you now?

My heart sank when I found out you died.

The last few years of your life were dark. Many of us were sad for your home
life, and to a large degree, that reflected in your work life.

I couldn't get rid of the thought in my head... what if you were not a true
Christian? What if you are in hell right now?

[Luke 16:19-31](https://biblia.com/bible/esv/luke/16/19-31) popped into my head.
And I realized, even if you are in hell, you would want to warn us all to live
in such a way as to escape hell.

For those of us who call ourselves Christians, you would warn us sternly to
persevere to the end. [Hebrews 3:14](https://biblia.com/bible/esv/hebrews/3/14).
[1 John 2:19](https://biblia.com/bible/esv/1-john/2/19).

For those of us who do not claim to know/**conhecer** Jesus, you would urge us
to take Jesus' free gift of salvation. You would explain to us that all humans
are sinful and deserve an eternity in hell for perverting God's name. But God,
rich in mercy, sent his Son, Jesus, to die for us, so that the wrath of God that
we deserve, would be imprinted on him, and his righteousness would be imprinted
on us.

You would jump up and down, and point to the finite age of the Universe, with an
exact beginning. You would point out the existence of matter, the laws of
physics, of math, of Euler's number, of the Golden Ratio, and of all the unique
properties of the elements in the periodic table that allow for a Universe to
sustain intelligent life. And if the Universe is finely tuned, it is finely
tuned by God.

You would point to the racism, sexual immorality, greed, and violence that is
plain for all to see. None could say with a straight face that evil does not
exist. And if objective evil and morality exist, they must be defined outside of
us, by God.

And you would ask us all to examine our hearts, and if we are honest with
ourselves, we would see that we have all been racist, sexually immoral, greedy,
and violent. We have all defamed the name of God.

And you would point us to Jesus, God in the flesh, who loves us, and nailed on
himself, on the cross, all the guilt and punishment that we deserve for defaming
God's name, to give those who hope in him, eternal life.

You would say to me, and urge me, to hold onto Jesus, because he holds onto me,
so as to not end up in hell.

I reached out to @sez11a. And I asked him straight up, "Do you think Nate is in
hell?" @sez11a rebuked me and affirmed his confidence in your salvation as I sat
broken. Many other Christians agreed, affirmed, and comforted me.

Nate, [Hebrews 6:9](https://biblia.com/bible/esv/hebrews/6/9) popped into my
head. And that verse gave me strength.

Later, I also remembered, all the long conversations we had together, of us
exulting in God's grace and mercy toward us undeserving sinners.

Yes, Nate, we all deserve hell, but Jesus died to save you, and me, and all who
hope in his grace.

Nate, when Jesus returns, and Heaven and Earth are merged together, and the
Earth is renewed, and we both run around with our new bodies, eating and
drinking... I want to work with you again on another open source project. How
about we spend a 100 years or a millenia on it? You do the frontend, I'll do the
backend. Sounds good?

Your friend,
Brian Chan