LPS-121736 Removes the swfobject library and wrapping AUI module
The AUI module definition was used to add the SWFOBJECT library to the
global scope as a side effect of the module loading.

For completeness, the archived chat application has been updated to
include the SWFObject library from an external CDN like jsdelivr.

This commit also removes a leftover `swfuplodader` AUI module that had
no proper backing anymore.