LPS-122265 Update @liferay/npm-bundler to v3.0.0-pre.1
This is just a routine update for hygiene with no
visible impact. We updated to v3.0.0-pre.0 yesterday in:

    https://issues.liferay.com/browse/LPS-122157

but after that we applied a small tweak:

    https://github.com/liferay/liferay-frontend-projects/pull/160

for the benefit of the Fragments Toolkit:

    https://github.com/liferay/generator-liferay-fragments

and cut a pre.1 release. Release notes here:

    https://github.com/liferay/liferay-frontend-projects/releases/tag/npm-bundler%2Fv3.0.1-pre.1

As a matter standard practice, we want to keep the versions in sync
everywhere, so let's do a small update in DXP as well even though it
won't have any visible effect beyond the version number.

Test plan: Deploy `frontend-js-recharts` and see it build and work (ie.
in content dashboard).