LPS-121729 Removes Open In MS Office items
IE11 will no longer be supported in Liferay DXP 7.4. That means that any
specific features that require that browser won't be supported either.

This commit removes the conditional addition of Open in MS Office items
to different menus or toolbars since the condition they require to work
(IE on Windows) are not to be satisfied in the next version.

This commit does not remove MS Office support, simply the UI that can
trigger the feature.