LPS-121753 Drop IE support from build in backportable way
Previously, we used to transpile all of our ".js" files to be ES5
compatible. That's a very old standard (5.1 was finalized in 2011), and
goes significantly beyond what we strictly needed to in order to meet
the requirements of our compatibility matrix.

In this commit, we not only drop support for IE, but we make our Babel
config very strictly match exactly what is in the compatibility matrix.
As an example of the gains, take frontend-js-web as an example, we see
this change strip off about 1,305 lines of code, and 69,752 characters
(pre-gzip):

Before:

    $ cloc build/node/packageRunBuild/resources

    -------------------------------------------------------------------------------
    Language                     files          blank        comment           code
    -------------------------------------------------------------------------------
    JavaScript                     139           1589           2345           8142

    $ wc build/node/packageRunBuild/**/*.js | tail -1
       12005   44520  580077 total

After:

    $ cloc build/node/packageRunBuild/resources

    -------------------------------------------------------------------------------
    Language                     files          blank        comment           code
    -------------------------------------------------------------------------------
    JavaScript                     139           1407           2529           6836

    $ wc build/node/packageRunBuild/**/*.js | tail -1
       10701   39260  510325 total

Here is a concrete example of a file, side_navigation.es.js:

- Before: https://gist.github.com/wincent/be6203a1e3032251a8974bcc633d3751
- After: https://gist.github.com/wincent/b2b073003ce91b9256e1b86bbb1b8ac2
- Diff: https://gist.github.com/wincent/408131bf38c070b10574e976930ad0aa

Believe it or not, this one is safe to backport to 7.3 because we have
implemented branch-specific behavior as described in this PR:

    https://github.com/liferay/liferay-frontend-projects/pull/214

The major version bump in `@liferay/npm-scripts` is because in a very
narrow _technical_ sense, it is a "breaking change" (ironically, only on
7.4). This is because it requires us to add one .babelrc.js for the
benefit of analytics-client-js where we continue to want IE support to
be in place, even though it lives on 7.4. The file will be harmless
no-op on 7.3, so nothing will break on it.

Release notes for `@liferay/npm-scripts` v34.0.0:

    https://github.com/liferay/liferay-frontend-projects/releases/tag/npm-scripts%2Fv34.0.0