LPS-121753 Don't use object concise methods for AUI component constructors
So I spent a bunch of time digging into why AUI doesn't like the object
concise method, looking at AUI source code, snooping with a debugger
etc, because for all intents and purposes, I couldn't see any real
difference between a method defined on an object using "longhand":

    {
        a: function a() {},
    }

and one using concise method syntax:

    {
        a() {}
    }

They have the same name, the same calling behavior, the same
enumerability etc. I was suspicious then, of AUI doing something deeply
weird deep down inside.

But it turns out that there actually _is_ a difference in calling
behavior, which ends up getting manifested in this case as:

    TypeError: Liferay.SearchContainer is not a constructor

I couldn't find an explanation of this on MDN (it is probably in the
spec somewhere though), but I did find this simple example on Stack
Overflow showing that this is expected behavior for concise methods:

    https://stackoverflow.com/a/45119651/2103996

With the benefit of hindsight, we can see that I could have figured
this out in five minutes instead of two hours. :facepalm:

Amazingly, there is exactly _one_ place in all of liferay-portal where
we actually pass a `constructor` into `A.Component.create`, so there is
exactly one place to fix. Found them by searching for `constructor` in
all files that contain the string `Component.create`; then did a broader
search for `constructor` in all ".js" and ".jsp" files that contain the
string `create`. It's possible something a bit obfuscated may have
slipped through, but I think this is a reasonable first pass.