LPS-121719 Use master version of svg4everybody.js
The project is pretty much unmaintained at this point: the last release
was about 3 years ago, although the `master` branch has commits as
recently as April this year (2020).

So, we take a snapshot of the source by grabbing the LICENSE.md and the
lib/svg4everybody.js source file as of current HEAD of `master`, and
basically concatenating them:

- https://raw.githubusercontent.com/jonathantneal/svg4everybody/8bd711049ba3faefb8285ca70c63a67efc35636c/LICENSE.md
- https://raw.githubusercontent.com/jonathantneal/svg4everybody/8bd711049ba3faefb8285ca70c63a67efc35636c/lib/svg4everybody.js

The project is simple, 0 dependencies, and 1 source file. We can ignore
the Gruntfile based UMD build that the project uses and just compile
this with `@liferay/npm-scripts`.

Changes made to the source:

-   Wrap whole thing in an IIFE and assign to `window.svg4everybody`.
-   Format source.
-   Fix lints:

        143:10  error  'svg4everybody' is defined but never used.  no-unused-vars
        153:6  error  'LEGACY_SUPPORT' is not defined.  no-undef
        187:13  error  'LEGACY_SUPPORT' is not defined.  no-undef
        245:9  error  'LEGACY_SUPPORT' is not defined.  no-undef
        302:9  error  Expected object keys to be in natural ascending order. 'id' should be before 'svg'.  sort-keys

    Note that we can just delete all the sections guarded by
    `LEGACY_SUPPORT` because we never used the legacy build.

Because the build is simpler now, we can strip out most of the
`build.gradle` file.

Test plan:

In portal-ext.properties, set:

    cdn.dynamic.resources.enabled=false

(This is the only way to inject svg4everybody, unless you are on IE,
which you no longer can be on `master`).

Load portal and see script successfully retrieved from:

    /o/frontend-js-svg4everybody-web/index.js

See transpiled source looks right:

    https://gist.github.com/wincent/5c61c838e0b8e3445e552c167cd9f2ef

And that `svg4everybody()` exists in the console.