LPS-122523 Removes dependency with 'metal'
It replaces some basic helpers from metal:
- Uses spread instead of object.mixin
- Uses typeof directly for type checks
- The `Disposable` class has been copied verbatim
- The `getUid` method has been ported over just as a simple counter

Additionally, it replaces async.nextTick with setTimeout(cb, 0). This
technically adds a min 50ms to those calls since it does not work as
a setImmediate. We should keep an eye on performance to see if we want
to bring back a traditional postMessage-based polyfill.