LPS-122956 Removes server-side soy rendering from SoyRenderer implementation
This basically means that the server will render an empty placeholder
and then components (if passed) will be hydrated client-side.

Additionally, we introduce a temporary workaround to reduce cumulative
layout shift. If a `__placeholder__` property is passed in the `context`
object, the renderer will use that instead of an empty placeholder so tag
implementors can choose to manually generate it temporarily.