LPS-124430 Saves tooltip trigger title on show events
Before, the change of `title` to `data-restore-title` happened based on the
component rendering and state, making it fail to execute under certain
conditions.

With this change, anytime the mouse moves over a tooltip trigger, its
`title` attribute is moved to a `data-restore-title` attribute avoiding
unnecessary rendering of the native tooltip.