LPS-123956 Hardcoding value to avoid exposing unnecessary API
Adding API + a dependency for this is too much. Stopping one module would
render the other one unsusable which is not desired. We just want to
remove the menu item if it is present, if not it can be safely ignored.

I originally proposed this solution to Alicia but after discussing it
further we decided to avoid the potential issue.