LPS-125168 Avoids unnecessary servlet request to retrieve Loader configuration
This commit removes at least 2 layers of indirection in the way we're configuring
the existing amd loader:
- Inlines the contents of /o/js_config
- Makes the output match the expected configuration from frontend-js-web/loader/config

Doing this we speed up initialization and avoid weird network requests.

Ideally, we should simply print the configuration within the loader initialization but
this is not currently possible without an important loader update. Since we're moving
towards Webpack5 and away from it, this sounds like a good middle ground until then