LPS-124287 npx yarn-deduplicate
Note that we are replacing webpack 5.4.0 in js-toolkit-core and npm-bundler by webpack 5.12.3
coming from the new version of npm-scripts. This is because semver ranges allow for it and we
don't want to have two webpacks in the build.

We could have released new versions of npm-bundler and js-toolkit-core pointing to ^5.11.1
explicitly but since it is not strictly necessary and since we are using them temporarily until
everything is switched to webpack 5, I don't want to make two new releases "for nothing".