LPS-125168 Caches AUI Bundle configuration output
Moving the work from the untracked servlet to the main thread resulted
in an 6% performance regression in Session Count detected in our automated
performance regression testing.

The net effect to the whole system should actually be a perf gain since we've removed
several components and unnecessary servlets which should reduce the invocation stack
on each request.

In any case, we're adding a caching layer here to reduce unnecessary overhead.