LPS-125478 Adding missing DDM structures
Empty document types coming from 7.3 have no dataDefinitionId associated
(0).

In 7.4 there is a unique index for the dataDefinitionId field
therefore we can't have more than 1 document type with the same value
(0) for a given group.

With this upgrade we create a structure for each of those document types
so that there are no document types left without one (apart from BASIC
DOCUMENT).

We also recreate the index in case it failed because the DB was already
in an incompatible state (more than 1 document type with the same
dataDefinitionId (0)).