LPS-125673 Fixes error when closing the sidebar and leaves a large blank space
This is a regression caused by changes to the LPS-124480 ticket that brings the state of the Sidebar to be internal to the component but the application needs this information to be able to resize the builder container. Just taking the state back into context but keeping the component decoupled from the context.