LPS-125168 Remove the cache and instead don't rely on JSONObjects to print the output
We have made some performance tests and this leads to a huge increase in performance:
from 12 ms to construct the JSONObject and 110 ms to print it, to 14 ms to print it
directly. That's a ~88% increase in performance.

Given that now, the total time to do the whole task has gone from ~166 ms to ~37 ms,
we don't need the cache any more. Also, adding a cache wouldn't benefit much, because
we would need to make the same operations we are doing now in addition to calculating
hashes so it could be even worse sometimes (as the operations we are performing are
almost instant as seen in the performance tests).

Just for the sake of reference, the only operation that could benefit from caching
right now would be obtaining the `resolvePath`, which takes ~21 ms. The rest are almost
insignificant (between 0 and 3 ms). Given that we will be removing the `resolvePath`
soon (webpack 5) I don't think it makes sense to make this more complex by re-adding a
cache.