LPS-125746 Modify frontend-js-loader-modules-extender API
We now have a new way to load JS modules (in addition to AMD loader): webpack.
Because of that, we need to cope with two types of OSGi modules too: the ones
built for AMD and the ones build for webpack. Thus, we add the notion of throwing
an UnsupportedOperationException when the NPMResolver is used for an OSGi bundle
not targeting AMD (but webpack instead).

This way we can make reactive code that, when calling NPMResolver, can catch the
UnsupportedOperationException and resort to the webpack layer instead to do its
duties.